// src/common/types.ts
export type ARxTransaction = {
  sign: () => Promise<Buffer>;
  size: number;
  uploadWithReceipt: (opts?: UploadOptions) => Promise<UploadReceipt>;
  upload(opts: UploadOptions & { getReceiptSignature: true }): Promise<UploadReceipt>;
  upload(opts?: UploadOptions): Promise<UploadResponse>;
  isValid(): Promise<boolean>;
  // fromRaw(rawTransaction: Buffer, ARxInstance: ARx): ARxTransaction;
} & DataItem;
export type ARxTransactonCtor = new (
  data: string | Uint8Array,
  ARx: Pick<ARx, "uploader" | "tokenConfig" | "arbundles">,
  opts?: ARxTransactionCreateOptions,
) => ARxTransaction;
// ...
export type UploadResponse = {
  // The ID of the transaction
  id: string;
  // The Arweave public key of the node that received the transaction
  public: string;
  // The signature of this receipt
  signature: string;
  // the maximum expected Arweave block height for transaction inclusion
  deadlineHeight: number;
  // List of validator signatures
  validatorSignatures: { address: string; signature: string }[];
  // The UNIX (MS precision) timestamp of when the node received the Tx.
  timestamp: number;
  // The receipt version
  version: "1.0.0";
  // Injected verification function (same as Utils/ARx.verifyReceipt)
  verify: () => Promise<boolean>;
};

export type UploadReceipt = /* Required<UploadResponse>; */ UploadResponse;
export type UploadReceiptData = Omit<UploadReceipt, "verify" | "validatorSignatures">;

export type FundResponse = {
  reward: string;
  target: string;
  quantity: string;
  id: string;
};
export type WithdrawalResponse = {
  tx_id: string;
  requested: number;
  fee: number;
  final: number;
};

export type CreateAndUploadOptions = DataItemCreateOptions & { upload?: UploadOptions };

export type UploadOptions = { paidBy?: string };

export enum UploadHeaders {
  PAID_BY = "x-arx-paid-by",
}

// // TS doesn't like string template literals it seems
// export enum manifestType {
//     paths = "arweave/paths"
// }

// export enum manifestVersion {
//     "0.1.0" = "0.1.0"
// }

export type ARxTransactionCreateOptions = DataItemCreateOptions & { dataIsRawTransaction?: boolean };